handlers: refactor with routers
